requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
boto3==1.34.130
botocore==1.34.130
certifi==2024.2.2
click==8.1.7
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.2
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.3
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.2
packaging==24.1
pillow==10.3.0
postgrest==0.16.8
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
Pygments==2.17.2
pyppeteer~=2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
realtime==1.0.6
rich==13.7.1
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.9.0
urllib3==1.26.19
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets
zipp==3.19.2

requests~=2.32.3
openai~=1.37.0